BlobReader and BlobWriter is at least working for simple test cases. More testing needed.

+++ src/dots/dots_internal.ss/src/Blob.cpp
@@ -69,7 +69,7 @@ namespace Internal
     boost::int64_t Blob::GetTypeId(const char *blob)
     {
         boost::int64_t val;
-        memcpy(&val, blob+sizeof(boost::int32_t), sizeof(boost::int32_t));
+        memcpy(&val, blob+sizeof(boost::int32_t), sizeof(boost::int64_t));
         return val;
 
     }
